refactor(evm-ir): simplify pass helpers from review
Truncate after terminals once before the CFG cleanup loop, shift
relocations with `retain_mut`, merge the two lazy-chain branches in
tail merging, rotate loop layout order in place, store outline ids as
`u32`, and fold the data scan into one loop.

+++ crates/codegen/src/backend/evm/ir/passes/reorder_pushes.rs
@@ -103,8 +103,8 @@ impl ReorderState {
         self.sequence.reserve(self.sequence.instructions.len());
         self.expressions.clear();
         let mut changed = false;
-        for index in 0..self.sequence.instructions.len() as u32 {
-            let inst = &self.sequence.instructions[index as usize];
+        for index in 0..self.sequence.instructions.len() {
+            let inst = &self.sequence.instructions[index];
             let swap1 = inst.as_stack_op() == Some(StackOp::Swap(1));
             if swap1
                 && let Some(pushed) = self.expressions.last()
@@ -135,7 +135,7 @@ impl ReorderState {
                 continue;
             }
 
-            let node = self.sequence.push(index);
+            let node = self.sequence.push(index as u32);
             update_expressions(&mut self.expressions, &self.sequence, node);
         }
         self.sequence.finish_into(instructions, changed);
